Adding notes
This final step is mostly to tie things together: it adds the
ability to add new notes and view existing ones. From here, you
should know enough to expand on the functionality: you could add
the ability to add a title (a schema:headline?) to the note, or
add other metadata such as the author (a schema:creator), or just
take the things you've learned now and build a different type of
web app altogether!

The relevant code in this commit is in `addNote`. Given a note (a
string), it initialises a new Subject with the note as its
contents (through schema:text) and the current time as
schema:dateCreated. Then, the <NotesList> component can use that
date to sort the user's notes by date, from newest to oldest.

So that's it for now! Study the code, start building your own web
apps (you can usually find Vocabularies for your use cases at
Linked Open Vocabularies [1]), and if you have questions, make
sure to ask them at the Solid Forum [2].

[1] https://lov.linkeddata.es/dataset/lov/
[2] https://forum.solidproject.org/